+++ test/lightning/credentials_test.exs
@@ -69,6 +69,118 @@ defmodule Lightning.CredentialsTest do
       assert Credentials.get_credential!(credential.id) == credential
     end
 
+    test "get_credential_by_project_credential/1 returns the credential with given project_credential id" do
+      refute Credentials.get_credential_by_project_credential(
+               Ecto.UUID.generate()
+             )
+
+      project_credential = insert(:project_credential)
+
+      credential =
+        Credentials.get_credential_by_project_credential(project_credential.id)
+
+      assert credential.id == project_credential.credential.id
+    end
+
+    test "create_credential/1 with valid data creates a credential" do
+      valid_attrs = %{
+        body: %{},
+        name: "some name",
+        user_id: user_fixture().id,
+        schema: "raw",
+        project_credentials: [
+          %{project_id: project_fixture().id}
+        ]
+      }
+
+      assert {:ok, %Credential{} = credential} =
+               Credentials.create_credential(valid_attrs)
+
+      assert credential.body == %{}
+      assert credential.name == "some name"
+
+      assert from(a in Audit.base_query(),
+               where: a.item_id == ^credential.id and a.event == "created"
+             )
+             |> Repo.one!(),
+             "Has exactly one 'created' event"
+    end
+
+    test "create_credential/1 with invalid data returns error changeset" do
+      assert {:error, %Ecto.Changeset{}} =
+               Credentials.create_credential(@invalid_attrs)
+    end
+
+    test "update_credential/2 with valid data updates the credential" do
+      user = user_fixture()
+
+      {:ok, %Lightning.Projects.Project{id: project_id}} =
+        Lightning.Projects.create_project(%{
+          name: "some-name",
+          project_users: [%{user_id: user.id}]
+        })
+
+      credential =
+        credential_fixture(
+          user_id: user.id,
+          project_credentials: [
+            %{project_id: project_id}
+          ]
+        )
+
+      original_project_credential =
+        Enum.at(credential.project_credentials, 0)
+        |> Map.from_struct()
+
+      new_project = project_fixture()
+
+      update_attrs = %{
+        body: %{},
+        name: "some updated name",
+        project_credentials: [
+          original_project_credential,
+          %{project_id: new_project.id}
+        ]
+      }
+
+      assert {:ok, %Credential{} = credential} =
+               Credentials.update_credential(credential, update_attrs)
+
+      assert credential.body == %{}
+      assert credential.name == "some updated name"
+
+      audit_events =
+        from(a in Audit.base_query(),
+          where: a.item_id == ^credential.id,
+          select: {a.event, type(a.changes, :map)}
+        )
+        |> Repo.all()
+
+      assert {"created", %{"after" => nil, "before" => nil}} in audit_events
+
+      assert {"updated",
+              %{
+                "before" => %{"name" => "some name"},
+                "after" => %{"name" => "some updated name"}
+              }} in audit_events
+
+      assert {"added_to_project",
+              %{
+                "before" => %{"project_id" => nil},
+                "after" => %{"project_id" => new_project.id}
+              }} in audit_events
+    end
+
+    test "update_credential/2 with invalid data returns error changeset" do
+      user = user_fixture()
+      credential = credential_fixture(user_id: user.id)
+
+      assert {:error, %Ecto.Changeset{}} =
+               Credentials.update_credential(credential, @invalid_attrs)
+
+      assert credential == Credentials.get_credential!(credential.id)
+    end
+
     test "delete_credential/1 deletes a credential and removes it from associated jobs and projects" do
       user = user_fixture()